fix debug dir, update diag output, don't delete tmp files during debug

+++ makeDVD/auto.sh
@@ -58,15 +58,16 @@ read -p "Oracle SSO Password: " ORACLE_PASSWORD; echo
 stty echo
 
 # login to oracle website first
+echo "LOGGING IN TO ORACLE SSO"
 curl --location-trusted -c /tmp/cookies -A "Mozilla/5.0" http://www.oracle.com/webapps/redirect/signon >/tmp/formfields
 getFormFields /tmp/formfields >/tmp/formx
-curl -vd @/tmp/formx -d ssousername="$ORACLE_USERNAME" -d password="$ORACLE_PASSWORD" --location-trusted -b /tmp/cookies -c /tmp/cookies -A "Mozilla/5.0" https://login.oracle.com/oam/server/sso/auth_cred_submit >$(dirname $0)/form_login_debug 2>&1
+curl -vd @/tmp/formx -d ssousername="$ORACLE_USERNAME" -d password="$ORACLE_PASSWORD" --location-trusted -b /tmp/cookies -c /tmp/cookies -A "Mozilla/5.0" https://login.oracle.com/oam/server/sso/auth_cred_submit >/tmp/form_login_debug 2>&1
 
 # download files from list
 cd $1
 for URL in "${URLS[@]}"; do
   FILE="$(basename $URL)"
-  echo "DOWNLOAD: /tmp/$FILE"
+  echo "DOWNLOADING: /tmp/$FILE"
   if [ ! -f /tmp/$FILE ]; then
     curl --location-trusted -b /tmp/cookies -c /tmp/cookies -A "Mozilla/5.0" -o /tmp/$FILE "$URL"
   fi
@@ -78,7 +79,10 @@ for URL in "${URLS[@]}"; do
   fi
 done
 
-rm /tmp/cookies
-rm /tmp/formx
-rm /tmp/formfields
+if [ ! -n "$DEBUG" ]; then
+  rm /tmp/cookies
+  rm /tmp/formx
+  rm /tmp/formfields
+  rm /tmp/form_login_debug
+fi
